# Speed-up the dipole magnetic field by ~25x