@santisoler 2022-11-10
Thanks for tackling this @leouieda. It was on my todo list! Merging since this is super clean, as always.